Update salon.py
To be more explicit, I edited this code so that float type and integer data types did not mix.  Python didn't seem to fault operations mixing these types, but it is better to keep things consistent.  Since the prices are actually a decimal (float) type, I changed the prices list to reflect the compatible data type.